Correct installation to only be in DAMLA venv
As the command `source activate DAMLA` opens a new shell any commands
that Docker needs to run in that shell need to be passed to RUN at the
same time. This correction ensures that the installation of packages
only happens in the DAMLA venv.

Additionally, as Bash is wanted for anytime source is used, have Docker
use it throughout by setting SHELL to be Bash.