Clean up the HOME area after install
Remove the src dir that mls was installed from. As this is a Docker
image, and so persistance should not be assumed between sessions, it is
reasonable to assume that any changes to the mls environment will get
dealt with through rebuilding the image and not through updating the
image each time.